bionicpro-auth/pyproject.toml
[project]
name = "bionicpro-auth"
version = "0.1.0"
description = "Auth app"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.4",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.1",
    "python-keycloak>=7.0.3",
    "redis>=7.1.0",
    "uvicorn[standard]>=0.40.0",
]

[dependency-groups]
dev = [
    "ruff>=0.15.0",
]
